rename file and add code to remove role

+++ database/migrations/2024_06_18_174043_data_seed_remove_analytics_resource_role.php
@@ -7,7 +7,13 @@ use Illuminate\Database\Migrations\Migration;
 return new class () extends Migration {
     public function up(): void
     {
-        DB::table('roles')->where('name', 'analytics.analytics_management')->delete();
+        $role_details = DB::table('roles')->where('name', 'analytics.analytics_management')->first();
+
+        if (! empty($role_details)) {
+            DB::table('model_has_roles')->where('role_id', $role_details->id)->delete();
+            DB::table('role_has_permissions')->where('role_id', $role_details->id)->delete();
+            DB::table('roles')->where('name', 'analytics.analytics_management')->delete();
+        }
     }
 
     public function down(): void
